requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pymongo==4.6.0
dnspython==2.4.2
django-cors-headers==4.3.1
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0
